Share guild WS URL builder, merge realtime twins
Extract buildGuildWsUrl (lib/wsUrl.ts) — the ws/wss URL-from-API-base
block was copy-pasted in four files. useRealtimeUpdates keeps its
SSR guard + fallback; useCollaboration keeps its memo; both now call
the shared builder.

useQueueRealtime and useCounterGroupRealtime were character-identical
except the URL segment and invalidators — one concept (content-free
WS change signal -> invalidate detail + list, refetch through gated
endpoints). They now share one implementation in useResourceRealtime.ts;
public hook names and signatures unchanged.

Refs #871
